requirements.txt
streamlit
pandas
numpy
scikit-learn
tensorflow-cpu
joblib